+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 ##Changelog
 
+###Flotilla 0.0.2 (0.0.0)
+
+- extend Ctx with cross handler functions
+
+
 ###Flotilla 0.0.1 (8.20.2014)
 
 - reforked, renamed as Flotilla
 - ini style configuration read into app environment
-- basic Flotilla interface for extension 
-
+- basic Flotilla interface for extension
+- basic Jingo templating
+- provisions for binary static or template Asset inclusion per engine
+ 
 
 ###Fleet 0.0 (7.22.2014)